--- a/parser.go
+++ b/parser.go
@@ -704,7 +704,7 @@ func Delete(data []byte, keys ...string) []byte {
 		newOffset = prevTok + 1
 	}
 
-	// We have to make a copy here if we don't want to mangle the original data, because byte arrays are
+	// We have to make a copy here if we don't want to mangle the original data, because byte slices are
 	// accessed by reference and not by value
 	dataCopy := make([]byte, len(data))
 	copy(dataCopy, data)